requirements.txt
# ----------------- Core -----------------
streamlit>=1.25.0           # Modern Streamlit UI
requests>=2.32.0            # HTTP requests for Serper API
jinja2>=3.1.3               # Email templates

# ----------------- LangChain & LLM -----------------
langchain==1.27.0           # LangChain framework
openai==1.31.0              # OpenAI API client

# ----------------- Database -----------------
pysqlite3-binary==0.4.3     # SQLite driver (faster, ensures thread safety)

# ----------------- Email & Utilities -----------------
email-validator==2.0.0       # Validate email addresses
python-dotenv==1.0.1         # Load environment variables from .env
tenacity==8.2.2              # Retry logic for network calls (optional but recommended)

# ----------------- Concurrency / Background Tasks -----------------
threadpoolctl==3.2.0         # Threadpool for background tasks
concurrent-log-handler==0.9.22  # Optional: better logging in multithreaded apps
